requirements.txt
# Albert Agile — dépendances
langgraph>=0.2
langchain>=0.3
langchain-ollama>=1.0
langchain-anthropic>=0.3
langchain-google-genai>=2.0
langchain-chroma>=1.0
langchain-core>=1.2
pydantic>=2.7
chromadb>=1.5
fastapi
uvicorn
python-dotenv
langserve
langgraph-checkpoint-sqlite
sse-starlette

# RAG fallback (si Ollama indisponible)
langchain-huggingface>=1.0
sentence-transformers>=2.0
